+++ src/PVOutput.Net/Requests/Modules/StatisticPeriodRequest.cs
@@ -9,13 +9,12 @@ namespace PVOutput.Net.Requests.Modules
         public DateTime FromDate { get; set; }
         public DateTime ToDate { get; set; }
 
-        public override IDictionary<string, object> GetUriPathParameters() => new Dictionary<string, object>
+        public override IDictionary<string, object> GetUriPathParameters()
         {
-            ["sid1"] = SystemId,
-            ["df"] = FormatHelper.GetDateAsString(FromDate),
-            ["dt"] = FormatHelper.GetDateAsString(ToDate),
-            ["c"] = IncludeConsumptionImport ? 1 : 0,
-            ["cdr"] = IncludeCreditDebit ? 1 : 0
-        };
+            var parameters = base.GetUriPathParameters();
+            parameters["df"] = FormatHelper.GetDateAsString(FromDate);
+            parameters["dt"] = FormatHelper.GetDateAsString(ToDate);
+            return parameters;
+        }
     }
 }