+++ src/PVOutput.Net/Responses/PVOutputArrayResponse.cs
@@ -17,5 +17,15 @@ namespace PVOutput.Net.Responses
         /// List of the actual <typeparamref name="TResponseContentType"/> values in the response.
         /// </summary>
         public IEnumerable<TResponseContentType> Values { get; internal set; }
+
+        /// <summary>
+        /// Compares the response to <paramref name="other"/> for base equivalence.
+        /// Api rate information is disgarded when comparing.
+        /// <strong>Please note that the contents of Values are not compared.</strong>
+        /// </summary>
+        /// <param name="other">Other response to compare.</param>
+        /// <returns>True if both responses are equivalent.</returns>
+        public override bool IsEquivalentTo(PVOutputBaseResponse other)
+            => base.IsEquivalentTo(other) && HasValues == ((PVOutputArrayResponse<TResponseContentType>)other).HasValues;
     }
 }