+++ CHANGELOG.md
@@ -3,11 +3,17 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-### Unreleased
+## Unreleased
 
 ### Added
 
 - Added a logo for the NuGet package [#38](https://github.com/pyrocumulus/pvoutput.net/pull/38)
+
+## Fixed
+
+- Corrected `SystemService` returning teams, estimates or extended properties with counts greater than 0, but with empty/null content
+- Removed dead code from certain code paths found through coverage testing
+- Increased code coverage to near 100%
 
 ## [0.8.0] - 2020-08-29